Themes changed from hex to word
Currently to pick a theme you must reference the CSS file to know that
red is 09, blue is 0d, or wait is it 0e?  You see my point.Changing
theme names from theme-base-09 to theme-base-red

Creating a new README which will double as a live demo.  It also shows
how I used javascript on a restructuredText document.